+++ src/mlp.c
@@ -44,7 +44,7 @@ MLP *mlp_init(size_t input, size_t hidden, size_t output) {
   return mlp;
 }
 
-void mlp_forward(MLP *mlp, double *input, double *output) {
+void mlp_forward(MLP *mlp, int *input, double *output) {
   double *hidden = (double *)malloc(mlp->hidden_nodes * sizeof(double));
   if(!hidden) {
     fprintf(stderr, "hidden activation malloc failed");
